Renames role to contact-role
